requirements.txt
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.2.1
django-extensions==2.2.9
djangorestframework==3.11.0
python-decouple==3.3
django-filter
markdown
